buffer: fix negative index for large buffers
Fix incorrect negative index results for large buffers.
Add tests for number, Buffer and string searches beyond 2 GiB.

Fixes: https://github.com/nodejs/node/issues/66294

+++ src/node_buffer.cc
@@ -1158,7 +1158,7 @@ void IndexOfString(const FunctionCallbackInfo<Value>& args) {
   }
 
   args.GetReturnValue().Set(result >= search_end ? -1
-                                                 : static_cast<int>(result));
+                                                 : static_cast<double>(result));
 }
 
 void IndexOfBuffer(const FunctionCallbackInfo<Value>& args) {
@@ -1246,14 +1246,14 @@ void IndexOfBuffer(const FunctionCallbackInfo<Value>& args) {
   }
 
   args.GetReturnValue().Set(result >= search_end ? -1
-                                                 : static_cast<int>(result));
+                                                 : static_cast<double>(result));
 }
 
-int32_t IndexOfNumberImpl(Local<Value> buffer_obj,
-                          const uint32_t needle,
-                          const int64_t offset_i64,
-                          const int64_t end_i64,
-                          const bool is_forward) {
+double IndexOfNumberImpl(Local<Value> buffer_obj,
+                         const uint32_t needle,
+                         const int64_t offset_i64,
+                         const int64_t end_i64,
+                         const bool is_forward) {
   ArrayBufferViewContents<uint8_t> buffer(buffer_obj);
   const uint8_t* buffer_data = buffer.data();
   const size_t buffer_length = buffer.length();
@@ -1276,7 +1276,7 @@ int32_t IndexOfNumberImpl(Local<Value> buffer_obj,
     ptr = nbytes::stringsearch::MemrchrFill(buffer_data, needle, backward_end);
   }
   const uint8_t* ptr_uint8 = static_cast<const uint8_t*>(ptr);
-  return ptr != nullptr ? static_cast<int32_t>(ptr_uint8 - buffer_data) : -1;
+  return ptr != nullptr ? static_cast<double>(ptr_uint8 - buffer_data) : -1;
 }
 
 void SlowIndexOfNumber(const FunctionCallbackInfo<Value>& args) {
@@ -1297,14 +1297,14 @@ void SlowIndexOfNumber(const FunctionCallbackInfo<Value>& args) {
       IndexOfNumberImpl(buffer_obj, needle, offset_i64, end_i64, is_forward));
 }
 
-int32_t FastIndexOfNumber(Local<Value>,
-                          Local<Value> buffer_obj,
-                          uint32_t needle,
-                          int64_t offset_i64,
-                          bool is_forward,
-                          int64_t end_i64,
-                          // NOLINTNEXTLINE(runtime/references)
-                          FastApiCallbackOptions& options) {
+double FastIndexOfNumber(Local<Value>,
+                         Local<Value> buffer_obj,
+                         uint32_t needle,
+                         int64_t offset_i64,
+                         bool is_forward,
+                         int64_t end_i64,
+                         // NOLINTNEXTLINE(runtime/references)
+                         FastApiCallbackOptions& options) {
   HandleScope scope(options.isolate);
   return IndexOfNumberImpl(buffer_obj, needle, offset_i64, end_i64, is_forward);
 }
